Allow prefix-matching and dynamic agent path len
This commit is done for supporting the customized highway merging env.
1. Prefix-matching. Environment agents with the same prefix now can be
processed by the same policy and stored in the same replay buffer. For
example, in merging, we have vehicles on the main road and merging lane,
and each types of vehicles are controlled by a corresponding policy.
2. Dynamic agent path length. In most multi-agent environments, there
are prefixed number of agents, and they are alive throughout the entire
episode. However, in highway driving, it is common for vehicles to come
in and out the scene. Thus, the episode length for each vehicle is
dynamic and they are not required to come in or disappear together.